requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pyodbc==5.3.0
python-dotenv==1.2.1
